new level 12: and-or
