+++ grammars/level14-Additions.lark
@@ -1,7 +1,6 @@
-// adds list access with numbers and brackets
+// adds the while loop
 
-?atom: NUMBER | var | list_access | unsupported_number
-list_access: var _LEFT_SQUARE_BRACKET (INT | random | var) _RIGHT_SQUARE_BRACKET
-change_list_item: var _LEFT_SQUARE_BRACKET (INT | var) _RIGHT_SQUARE_BRACKET _SPACE _EQUALS _SPACE (var | textwithoutspaces)
-list_access_var: var _SPACE _EQUALS _SPACE var _LEFT_SQUARE_BRACKET (INT | random | var) _RIGHT_SQUARE_BRACKET
-assign_list: var _SPACE _IS _SPACE _LEFT_SQUARE_BRACKET (quoted_text | NUMBER) (_COMMA _SPACE? (quoted_text | NUMBER))+ _RIGHT_SQUARE_BRACKET
+command: print | ifs elses? | ask | for_loop | while_loop| assign_list | list_access_var | assign | comment
+while_loop: _WHILE _SPACE (condition) _EOL (_SPACE command) (_EOL _SPACE command)* _EOL _END_BLOCK
+
+
